+++ CHANGES.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Facial recognition using CompreFace
 - Interface to manage Traing Images on CompreFace server
 - API management endpoint and interface
+- Profile photos 
 
 ### Changed
 - Updated API calls to reflect appropriate request types